utils/logs.py: Logging exceptions via HTTP formats traceback into message string;
Previously the monitor.py script would not log the traceback logged for errors in controller. Since the logrecord was converted to a json object for transport, the exc_info attribute could not be read on the monitor side to get a traceback string. Now the utils.RemoteHandler class formats traceback string into the message before sending it via HTTP.